backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
PyMuPDF==1.24.13
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.3
